+++ src/schema_first/openapi/schemas/v3_1_1/external_docs_schema.py
@@ -4,15 +4,10 @@ from marshmallow import validate
 from ..base import BaseSchema
 from ..constants import RE_SERVER_URL
 from ..fields import DESCRIPTION_FIELD
-from .server_variable_object_schema import ServerVariableObjectSchema
 
 
-class ServerSchema(BaseSchema):
+class ExternalDocsSchema(BaseSchema):
     url = fields.String(
         required=True, validate=[validate.Regexp(RE_SERVER_URL), validate.Length(min=1)]
     )
     description = DESCRIPTION_FIELD
-    name = fields.String(validate=[validate.Length(min=1)])
-    variables = fields.Dict(
-        keys=fields.String(), values=fields.Nested(ServerVariableObjectSchema, required=True)
-    )