+++ tests/objects/server_object/conftest.py
@@ -11,6 +11,7 @@ def fx_server_object_full(fx_server_object_required) -> dict:
     return {
         **fx_server_object_required,
         'description': 'The production API server',
+        'name': 'test',
         'variables': {
             'username': {
                 'default': 'demo',